Split data/value.rs single file module into multiple files
